code style format: clang-format applied to src/a*/*
This patch contains the .clang-format file, which defines the rules for formatting the code base. The file format-file.py is a wrapper around clang-format which applies additional custom rules not present (yet) in clang-format, which should be used for systemd.

The formatting rules in this patch have been applied to all files in src/a*/*

see also last pr here: #11690