hwdb: attempt to fix the messy vendor/product names
Some of the names contains just too much garbage to be actually
human-readable:

  ID_VENDOR_FROM_DATABASE=Qualcomm Inc
  ID_MODEL_FROM_DATABASE=AR9285 Wireless Network Adapter (PCI-Express) (AW-NB037H 802.11bgn Wireless Half-size Mini PCIe Card [AR9002WB-1NGCD])

There's just too much clutter there. E.g. the parts about the device
class or the bus are obviously redundant. The form factor, supported
frequencies, etc. are out of place.

What should the user be seeing in order to identify the hardware is:

  ID_VENDOR_FROM_DATABASE=Qualcomm
  ID_MODEL_FROM_DATABASE=AR9285 (AW-NB037H)

This upset NetworkManager users to the point it ended up patching things
around in libnm. Other tools probably suffer from a similar problem.

It would be really useful if the udev properties actually contained
high quality human-readable product and vendor names.

This commit patches the hwdb in a manner similar to what libnm does.
My python is probably horrible -- sorry. For now it mostly exists to:

1.) Illustrate the point
2.) See if this actually also bothers others
3.) Kick off a discussion about what would be the best way to tackle
    the problem.

Perhaps the real fix belongs to the upstream id databases; I haven't
tried sending it that way yet. I also don't know how to maintain the
reasonableness of the strings.

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>