logind: Allow any fbdev device to be seat master when KMS is disabled
When 'nomodeset' is specified, there's no DRM driver to take over from
efifb. This means no device will be marked as a seat master, so gdm will
never find a sufficiently active seat to start on.

I'm not aware of an especially good way to detect this through a proper
kernel API, so check for the word 'nomodeset' on the command line and
allow fbdev devices to be seat masters if found.

See also: https://bugzilla.redhat.com/show_bug.cgi?id=1683197