resolved: bind socket to the interface
When using a DNS server associated to a link, make sure to use this link to send the request.
IP_UNICAST_IF is overridden by the configured routes, and therefore doesn't guarantee that the request will go on the desired interface.
To use SO_BINDTODEVICE we need to keep the capability CAP_NET_RAW