Add git submodule for libbpf
The choice of having the library as a git submodule allows us to iterate faster
in bringing eBPF capabilities due to the package for Red Hat does not exist yet.

Github mirror of of bpf-next linux tree: https://github.com/libbpf/libbpf

The package for Debian can be found in https://packages.debian.org/sid/libbpf-dev

Meanwhile, there is an ongoing discussion for .rpm:
https://lore.kernel.org/bpf/CAADnVQKvRz-EMPqB5oMiG=6N1-QwAhMGnpnQsgSyQuUcrA781A@mail.gmail.com/T/#m89f1546f20062d39a27fc7d7a12b4af5b6325c2a
Once there is an availible RPM we'll replace the git submodule with a proper
external dependency.