pid1: when loading dropins, use only the main unit name, not aliases
We would sometimes also load dropins for the alias names, but not always.
The list of names to look at is passed as a set. In systemctl cat
we'd always pass a set with exactly one name.

In systemd proper, the situation is more complicated. We'd look for dropins
under any name known for the unit. This means that the results were inconsistent:
if a unit was first accessed under its primary name, systemd wouldn't know
of any aliases, so it'd only load the dropins for the main name. But if
the unit was first accessed under one of the aliases, systemd would know
some a bigger set of names, and possibly load dropins also from other paths.
This is related to the basic design rule that systemd does *not* load all the
configuration from disk, but only loads the unit files as they are requested.

This means that for the results to be consistent, i.e. to have the same set
of load paths for dropins, we can only look for dropins under the final name,
not under any aliases.

(I considered trying to detect aliases, e.g. by looking for the [Install] section
when loading units, but that'd be very frail, and could never cover all cases,
since the user is allowed to create an alias through a symlink without adding
anything to [Install]. So as long as we keep the decision to not scan the full
file system for aliases, the solution in this patch seems reasonable.)

A clarification is added to the docs.

Fixes #11972.