socket: Only flush accepts on listening sockets
Calling accept() on sockets which are not able to listen and accept
(.e.g. datagram sockets) may return EOPNOTSUPP, which is assumed to be
a transient error by flush_accept and retried ad infinitum.

Only try to flush pending connection requests on sockets which are able
to accept connections.

Fixes #12335