man: alternative path for resolved.conf
On some systems the default `resolved.conf` can be in `/lib/systemd/`. This commit adds the alternative path directly into the offline manual, because whoever is configuring the network may not yet have Internet access to find the file when it is not under `/usr/lib/systemd/`. Mitigates issue #8359 (GitHub).